Consolidate provider code and rename package
